refactor: 톡픽 DTO 내 공통 필드 클래스에 @Valid 추가 및 네이밍 개선
